Introduce run-time language switching
This incorporates all present translations into one build. Instead of a
symlink to the compile-time language, a switch is generated for every
translation module. This switch contains all the functions of the
original(s), but they take a Language as an extra (preceeding) argument,
delegating the translation itself to the corresponding module for that
language.

Migration of apps is needed, instructions in README.md.

NOTE: This breaks the currently implemented behaviour of "one language
per build".